services/ml/requirements.txt
fastapi
uvicorn[standard]
pydantic
onnxruntime
scikit-learn
numpy
python-dotenv
